+++ EmmaSharp/Models/Members/MembersAdd.cs
@@ -1,12 +1,10 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
 
 namespace EmmaSharp.Models.Members
 {
     public class MembersAdd
     {
         [JsonProperty("import_id")]
-        public int ImportId { get; set; }
+        public long ImportId { get; set; }
     }
 }